first pass moving daemon and transaction code to new crate
