+++ cfdp-core/src/lib.rs
@@ -1,8 +1,7 @@
 pub mod daemon;
 pub mod filestore;
 pub mod pdu;
-pub(crate) mod segments;
-pub(crate) mod timer;
+
 pub mod transaction;
 pub mod transport;
 // Re-exported for convenience
