+++ cfdp-core/src/transaction.rs
@@ -1,11 +1,7 @@
 mod config;
 mod error;
-mod recv;
-mod send;
 
 pub use self::{
     config::{Metadata, TransactionConfig, TransactionID, TransactionState},
     error::TransactionError,
-    recv::RecvTransaction,
-    send::SendTransaction,
 };
