+++ cfdp-daemon/src/segments.rs
